More random things

+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/fluid/tags/FluidTemperature.kt
@@ -5,4 +5,4 @@ import io.github.pylonmc.pylon.core.fluid.PylonFluidTag
 /**
  * Temperature in celsius, the superior unit of measurement
  */
-data class FluidTemperature(val temperature: Int) : PylonFluidTag
+data class FluidTemperature(val temperature: Double) : PylonFluidTag
